+++ Source Code/app/src/main/java/com/siddydevelops/newsdaily/fragments/EntertainmentFragment.java
@@ -1,4 +1,4 @@
-package com.siddydevelops.newsdaily;
+package com.siddydevelops.newsdaily.fragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,6 +10,12 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.siddydevelops.newsdaily.Adapter;
+import com.siddydevelops.newsdaily.ApiUtilities;
+import com.siddydevelops.newsdaily.ModelClass;
+import com.siddydevelops.newsdaily.R;
+import com.siddydevelops.newsdaily.mainNews;
 
 import java.util.ArrayList;
